GN-168 Refactor backToTop. Debug issue with Footer items left in main component files
