+++ src/Main/Main.module.scss
@@ -1,6 +1,6 @@
 @supports (position: sticky) {
   .main {
-    border: 5px solid red;
+    padding-bottom: 4rem;
     position: relative;
   }
 }
