+++ src/Main/index.js
@@ -1,1 +1,1 @@
-export { default } from "./Footer";
+export { default } from "./Main";
